made some refactoring
ok, maybe this will work finally so I can test a matrix approach
