+++ SideMenuExampleUITests/SideMenuExampleUITests.swift
@@ -1,6 +1,6 @@
 //
-//  SideMenu_ExUITestsLaunchTests.swift
-//  SideMenu_ExUITests
+//  SideMenuExampleUITests.swift
+//  SideMenuExampleUITests
 //
 //  Created by Derson Productions, LLC on 2024-AUG-03.
 //
@@ -32,26 +32,35 @@
 
 import XCTest
 
-final class SideMenu_ExUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
+final class SideMenuExampleUITests: XCTestCase {
 
     override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+
+        // In UI tests it is usually best to stop immediately when a failure occurs.
         continueAfterFailure = false
+
+        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
     }
 
-    func testLaunch() throws {
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() throws {
+        // UI tests must launch the application that they test.
         let app = XCUIApplication()
         app.launch()
 
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
 
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
+    func testLaunchPerformance() throws {
+        if #available(macOS 10.15, iOS 13.0, tvOS 13.0, watchOS 7.0, *) {
+            // This measures how long it takes to launch your application.
+            measure(metrics: [XCTApplicationLaunchMetric()]) {
+                XCUIApplication().launch()
+            }
+        }
     }
 }
